:recycle: REFACTOR: 서비스 평가 항목 저장 기능 수정
