+++ slash/src/main/java/project/slash/contract/exception/ContractErrorCode.java
@@ -9,8 +9,7 @@ import project.slash.common.exception.ErrorCode;
 @RequiredArgsConstructor
 @Getter
 public enum ContractErrorCode implements ErrorCode {
-	NOT_FOUND_CONTRACT(HttpStatus.BAD_REQUEST, "존재하지 않는 계약입니다."),
-	NOT_FOUND_ITEMS(HttpStatus.BAD_REQUEST, "올바른 서비스 평가 항목이 아닙니다.");
+	NOT_FOUND_CONTRACT(HttpStatus.BAD_REQUEST, "존재하지 않는 계약입니다.");
 
 	private final HttpStatus httpStatus;
 	private final String message;
