+++ slash/src/main/java/project/slash/evaluationitem/controller/EvaluationItemController.java
@@ -1,7 +1,5 @@
-package project.slash.contract.controller;
+package project.slash.evaluationitem.controller;
 
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,9 +7,8 @@ import org.springframework.web.bind.annotation.RestController;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import project.slash.common.response.BaseResponse;
-import project.slash.contract.dto.request.CreateDetailDto;
-import project.slash.contract.dto.response.EvaluationItemDetailDto;
-import project.slash.contract.service.EvaluationItemService;
+import project.slash.evaluationitem.dto.request.CreateEvaluationItemDto;
+import project.slash.evaluationitem.service.EvaluationItemService;
 
 @RequiredArgsConstructor
 @RestController
@@ -21,12 +18,12 @@ public class EvaluationItemController {
 	/**
 	 * 서비스 항목 세부 내용 설정 메서드입니다.
 	 *
-	 * @param detailDto 세부 내용 정보
+	 * @param createEvaluationItemDto 세부 내용 정보
 	 * @return 성공 여부
 	 */
 	@PostMapping("/detail")
-	public BaseResponse<Void> createDetail(@RequestBody @Valid CreateDetailDto detailDto) {
-		evaluationItemService.createDetail(detailDto);
+	public BaseResponse<Void> createEvaluationItem(@RequestBody @Valid CreateEvaluationItemDto createEvaluationItemDto) {
+		evaluationItemService.createEvaluationItem(createEvaluationItemDto);
 
 		return BaseResponse.ok();
 	}
