+++ slash/src/main/java/project/slash/evaluationitem/dto/request/CreateEvaluationItemDto.java
@@ -1,4 +1,4 @@
-package project.slash.contract.dto.request;
+package project.slash.evaluationitem.dto.request;
 
 import java.util.List;
 
@@ -6,13 +6,16 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
-import project.slash.contract.dto.GradeDto;
-import project.slash.contract.dto.TaskTypeDto;
+import project.slash.evaluationitem.dto.TaskTypeDto;
+import project.slash.evaluationitem.dto.ServiceTargetDto;
 
 @Getter
-public class CreateDetailDto {
-	@NotNull(message = "평가항목은 필수입니다.")
-	private Long categoryId;
+public class CreateEvaluationItemDto {
+	@NotNull(message = "계약 아이디는 필수입니다.")
+	private Long contractId;
+
+	@NotBlank(message = "서비스 평가 항목은 필수입니다.")
+	private String category;
 
 	@NotNull(message = "가중치는 필수입니다.")
 	private int weight;
@@ -34,5 +37,5 @@ public class CreateDetailDto {
 
 	@Valid
 	@NotNull(message = "서비스 목표는 필수입니다.")
-	private List<GradeDto> serviceTargets;
+	private List<ServiceTargetDto> serviceTargets;
 }
