+++ slash/src/main/java/project/slash/evaluationitem/repository/EvaluationItemRepository.java
@@ -1,10 +1,8 @@
-package project.slash.contract.repository.evaluationItem;
-
-import java.util.List;
+package project.slash.evaluationitem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import project.slash.contract.model.EvaluationItem;
+import project.slash.evaluationitem.model.EvaluationItem;
 
 public interface EvaluationItemRepository extends JpaRepository<EvaluationItem, Long>, EvaluationItemRepositoryCustom {
 }
