+++ slash/src/main/java/project/slash/evaluationitem/repository/ServiceTargetRepository.java
@@ -1,10 +1,8 @@
-package project.slash.contract.repository;
-
-import java.util.List;
+package project.slash.evaluationitem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import project.slash.contract.model.ServiceTarget;
+import project.slash.evaluationitem.model.ServiceTarget;
 
 public interface ServiceTargetRepository extends JpaRepository<ServiceTarget, Long> {
 }
